added html template
