+++ frontend/src/index.html
@@ -5,10 +5,10 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0, shrink-to-fit=no">
   <title>SDS Resources</title>
-  <link rel="SHORTCUT ICON" href="8fce6cce852b22a7a30d.svg" type="image/x-icon">
+  <link rel="SHORTCUT ICON" href="./assets/favicon.svg" type="image/x-icon">
   <script id="load-msal" src="https://alcdn.msauth.net/browser/2.15.0/js/msal-browser.js" integrity="sha384-dFzMiVGB5HpWZ+5w5VSif6jhWfNeplSw9ACYmQKZcY2azuT9kCxVWVI9HyfGdkHV" crossorigin="anonymous"></script>
   <script type="text/javascript">//if (typeof Msal === 'undefined') document.write(unescape("%3Cscript src='https://alcdn.msftauth.net/browser/2.15.0/js/msal-browser.js' type='text/javascript' crossorigin='anonymous' %3E%3C/script%3E"));</script>
-<link rel="icon" href="favicon.svg"><script defer src="runtime.c6d5c2e0032eaf512bf9.js"></script><script defer src="main.685ad97d1f212009d78a.js"></script></head>
+</head>
 
 <body>
   <div id="App">
@@ -16,7 +16,7 @@
     <div id="AppNavbar">
       <div id="AppNavbarLogo">
         <a href="http://sds.ubd.edu.bn/">
-          <img src="ef4bac32a69aec712b0d.png" alt="School of Digital Science">
+          <img src="./assets/logo.png" alt="School of Digital Science">
         </a>
       </div>
       <div id="AppNavbarNavs">
