--- frontend/src/stylesheets/box.scss
@@ -1,5 +1,0 @@
-$body-color: red;
-
-.box-component {
-  color: $body-color;
-}
