+++ frontend/src/stylesheets/components/app.scss
@@ -1,0 +1,7 @@
+#App {
+  margin: 0;
+  font-family: sans-serif;
+  .hidden {
+    display: none;
+  }
+}
