+++ frontend/src/stylesheets/components/app/main-container.scss
@@ -1,0 +1,23 @@
+#App {
+  #AppMainContainer {
+    transition: margin-left 0.2s ease-out;
+    margin-left: 20px;
+    #AppPage{
+      #AppPageTitle {
+        margin: 20px 0 0 20px;
+      }
+      #AppPageRemarks {
+        margin: 5px 0 0 20px;
+      }
+      #AppPageContent {
+        margin: 0 0 0 20px;
+        button {
+          margin: 0 ;
+        }
+        h3 {
+          margin: 20px 0 10px;
+        }
+      }
+    }
+  }
+}
