+++ frontend/src/stylesheets/components/app/message-box.scss
@@ -1,0 +1,22 @@
+$color_yellow-light: #ffdd88cc;
+
+#App {
+	#AppMessageBox {
+		z-index: 1;
+		display: block;
+		position:fixed;
+		bottom: 0;
+		width:100%;
+
+		text-align: center;
+
+		p,h3 {
+			display:block;
+
+			margin: 0;
+			padding: 7px 0;
+
+			background-color: $color_yellow-light;
+		}
+	}
+}
