+++ frontend/src/stylesheets/components/app/navbar.scss
@@ -1,0 +1,104 @@
+$color_navbar-bg: linear-gradient(105deg, rgba(14,60,110,1) 0%, rgba(14,60,110,1) 55%, rgba(0,212,255,1) 100%);
+$color_navbar-bg-filter: progid:DXImageTransform.Microsoft.gradient(startColorstr="#0e3c6e",endColorstr="#00d4ff",GradientType=1);
+$color_navbar-bg-plain: rgb(14,60,110);
+$color_yellow-darkest: #ea0;
+$color_yellow-dark: #f0c000;
+$color_yellow: #ffcb00;
+
+@mixin nav-link-animation {
+  transition-property: text-decoration-color, color;
+  transition-duration: 0.2s;
+  transition-timing-function: ease-out;
+}
+
+@mixin button-animation {
+  transition-property: background-color, color, box-shadow, top, left;
+  transition-duration: 0.2s;
+  transition-timing-function: ease-in-out;
+}
+
+#App {
+  #AppNavbarFiller { height: 55px; }
+  #AppNavbar {
+    z-index: 1;
+    position: fixed; display: flex;
+    left:0; top:0;
+    width: 100%; height: 55px;
+    
+    flex-direction: row;
+    justify-content: space-between; align-items: center;
+
+    background: $color_navbar-bg;
+    filter: $color_navbar-bg-filter;
+
+    box-shadow: 0 8px 16px 0 rgba(0,0,0,0.3), 0 6px 20px 0 rgba(0,0,0,0.29);
+
+    #AppNavbarLogo {
+      flex-grow: 1;
+      img { margin: 9px 0 6px 9px; }
+    }
+
+    #NavbarNavs {
+      display:flex;
+      margin: 0 2px;
+
+      flex-direction: row; flex-grow: 1;
+      justify-content: space-around; align-items: flex-end;
+
+      .nav {
+        &.hidden {display: none;}
+        .link {
+          display: table-cell;
+          height: 55px; padding:0 4px;
+
+          vertical-align: middle; text-align: center;
+
+          color: white;
+
+          text-decoration: underline rgb(14,60,110,0);
+          text-underline-offset: 4px;
+
+          @include nav-link-animation();
+
+          &:hover, &.current {
+            color: $color_yellow;
+            text-decoration: underline $color_yellow;
+          }
+        }
+      }
+    }
+  }
+  #AppNavbarAuth {
+    flex-grow: 200;
+    text-align: right;
+    margin-right: 10px;
+
+    #AppNavbarAuthButton {
+      position:relative;
+      top: 0px; left: 0px;
+      padding: 5px; border: none; border-radius: 5px;
+      
+      box-shadow: 3px 6px 4px 0 rgba(0,0,0,1);
+      background-color: $color_yellow;
+      color: black;
+
+      cursor: pointer;
+
+      @include button-animation();
+      &:hover {
+        position:relative;
+        top: 2px; left: 1px;
+
+        box-shadow: 1.5px 3px 2px 0 rgba(0,0,0,1);
+        background-color: $color_yellow-dark;
+      }
+      &.clicked, &:active {
+        position:relative;
+        top: 4px; left: 2px;
+        
+        box-shadow: 0px 0px 0px 0 rgba(0,0,0,1);
+        background-color: $color_yellow-darkest;
+      }
+    }
+  }
+}
